cli: common grid create/update options, payload parameters
